requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pdfplumber==0.11.6
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
aiosqlite==0.21.0
httpx==0.28.1
PyJWT==2.10.1
slowapi==0.1.9
